feat: add router and createRoot
